\name{defmacro}

\alias{defmacro}

\title{Internal function}

\description{Internal function
}

\usage{
defmacro(\dots, expr)}

\arguments{
  \item{\dots}{further arguments passed to or from other methods} 
  \item{expr}{a data frame with \emph{n} rows and \emph{p} columns, i.e. a contingency table}
}

\value{
}

\references{
}

\author{Jérémy Mazet, François Husson \email{François.Husson@agrocampus-rennes.fr}}

\keyword{multivariate}
