\name{Factocatdes}

\alias{Factocatdes}

\title{GUI for description of categories}

\description{GUI for description of categories
}

\usage{
Factocatdes()}

\arguments{
}

\value{
}

\references{
}

\author{Jérémy Mazet, François Husson \email{François.Husson@agrocampus-rennes.fr}}

\keyword{multivariate}
