\name{FactoPCA}
\alias{FactoPCA}

\title{
Perform PCA with the GUI
}
\description{
Perform PCA with the GUI
}
\usage{
FactoPCA()
}

\author{
F. Husson}

\examples{
\dontrun{
require(FactoMineR)
data(decathlon)
FactoPCA()
}}

\keyword{multivariate}
\keyword{print}
\keyword{dplot}
