\name{RcmdrPlugin.FactoMineR-internal}
\title{Internal RcmdrPlugin.FactoMineR objects}
\alias{defmacro}
\alias{FactoAFDM}
\alias{FactoCA}
\alias{Factocatdes}
\alias{FactoDMFA}
\alias{FactoGPA}
\alias{FactoHCPC}
\alias{FactoHMFA}
\alias{FactoMCA}
\alias{FactoMFA}
\alias{FactoPCA}
\alias{Factoprefpls}
\alias{readDataSetFacto}
\description{Internal RcmdrPlugin.FactoMineR objects.}
\details{These are not to be called by the user.}
\keyword{internal}
