\name{RcmdrPlugin.FactoMineR-package}
\alias{RcmdrPlugin.FactoMineR-package}
\alias{RcmdrPlugin.FactoMineR}
\docType{package}
\title{
RcmdrPlugin.FactoMineR package
}
\description{
RcmdrPlugin.FactoMineR package
}
\details{
RcmdrPlugin.FactoMineR gives a plugin to easier use the FactoMineR package. The package Factoshiny should be preferred at RcmdrPlugin.FactoMineR.

}
\author{
Francois Husson
Maintainer: husson <husson@agrocampus-ouest>
}
\seealso{
The FactoMineR package.
}
\keyword{ package }
