\name{RcmdrPlugin.FactoMineR}
\alias{RcmdrPlugin.FactoMineR}
\alias{RcmdrPlugin.FactoMineR}
\docType{package}
\title{
Graphical User Interface for FactoMineR}
\description{
FactoMineR "plug-in"
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.FactoMineR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-08-10\cr
License: \tab GPL version 2 or later\cr
}

}
\author{Francois Husson, Julie Josse, Sebastien Le \email{husson@agrocampus-rennes.fr}}

\keyword{ package }
